Move some classes in new packages - io.element.android.libraries.mediaviewer.impl.datasource - package io.element.android.libraries.mediaviewer.impl.model
